requirements.txt
# Core ML libraries
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0

# Evaluation metrics
rouge-score>=0.1.2

# Web scraping and HTTP
requests>=2.28.0
beautifulsoup4>=4.11.0

# Configuration
PyYAML>=6.0

# Optional: for better performance
accelerate>=0.20.0

# Required for some models (PEGASUS, etc.)
protobuf>=3.20.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
